fix(tui): session hub build errors (Ratio u32, borrow of route state)
- session_hub: Constraint::Ratio takes u32 in ratatui 0.30, not u16
- app_hub: build overview sections before mutably borrowing route state
  so the hub_session_item projection can borrow self immutably

+++ crates/agena-tui-session/src/session_hub.rs
@@ -274,12 +274,12 @@ fn split_section_areas(area: Rect, sections: &[SessionHubSection]) -> Vec<Rect> 
     }
     let total = sections
         .iter()
-        .map(|section| section.items.len().max(1) as u16)
-        .sum::<u16>()
+        .map(|section| section.items.len().max(1) as u32)
+        .sum::<u32>()
         .max(1);
     let constraints = sections
         .iter()
-        .map(|section| Constraint::Ratio(section.items.len().max(1) as u16, total))
+        .map(|section| Constraint::Ratio(section.items.len().max(1) as u32, total))
         .collect::<Vec<_>>();
     Layout::default()
         .direction(Direction::Vertical)
